+++ CHANGELOG.md
@@ -15,5 +15,6 @@
 - [feat: optimize pom dependency and demo](https://github.com/Tencent/spring-cloud-tencent/pull/71)
 - [feat:upgrade polaris version to 1.3.0.](https://github.com/Tencent/spring-cloud-tencent/pull/72)
 - [feat:rollback init polaris-context to application phase](https://github.com/Tencent/spring-cloud-tencent/pull/73)
+- [feat:add custom label resolver spi for rate limit](https://github.com/Tencent/spring-cloud-tencent/pull/74)